+++ myaccount/tpl/myaccount.tpl.php
@@ -71,7 +71,7 @@ print '
 	            <div class="caption-subject font-green-sharp bold uppercase">'.$langs->trans("Organization").' <small>('.$langs->trans("Code").' '.dolPrintHTML($mythirdpartyaccount->code_client).')</small></div>
 	          </div>
 	          <div class="portlet-body">
-
+				<!-- form for company profile -->
 	            <form action="'.$_SERVER["PHP_SELF"].'" method="post" name="formsoc">
 	            <input type="hidden" name="token" value="'.newToken().'">
 				<input type="hidden" name="action" value="updatemythirdpartyaccount">
@@ -81,27 +81,27 @@ print '
 
 	                <div class="form-group">
 	                  <label>'.$langs->trans("NameOfCompany").'</label>
-	                  <input type="text" class="form-control" placeholder="'.$langs->trans("NameOfYourOrganization").'" value="'.$mythirdpartyaccount->name.'" name="orgName">
+	                  <input type="text" class="form-control" placeholder="'.$langs->trans("NameOfYourOrganization").'" value="'.$mythirdpartyaccount->name.'" name="orgName" spellcheck="false">
 	                </div>
 
 	                <div class="form-group">
 	                  <label>'.$langs->trans("AddressLine").'</label>
-	                  <textarea class="form-control" placeholder="'.$langs->trans("HouseNumberAndStreet").'" name="address">'.$mythirdpartyaccount->address.'</textarea>
+	                  <textarea class="form-control" placeholder="'.$langs->trans("HouseNumberAndStreet").'" name="address" spellcheck="false" >'.$mythirdpartyaccount->address.'</textarea>
 	                </div>
 	                <div class="form-group">
 	                  <label>'.$langs->trans("Town").'</label>
-	                  <input type="text" class="form-control" value="'.$mythirdpartyaccount->town.'" name="town">
+	                  <input type="text" class="form-control" value="'.$mythirdpartyaccount->town.'" name="town" spellcheck="false">
 	                </div>
 	                <div class="form-group">
 	                  <label>'.$langs->trans("Zip").'</label>
-	                  <input type="text" class="form-control input-small" value="'.$mythirdpartyaccount->zip.'" name="zip">
+	                  <input type="text" class="form-control input-small" value="'.$mythirdpartyaccount->zip.'" name="zip" spellcheck="false">
 	                </div>
 	                <div class="form-group">
 	                  <label>'.$langs->trans("StateOrCounty").'</label>
-	                  <input type="text" class="form-control" name="stateorcounty" value="">
+	                  <input type="text" class="form-control" name="stateorcounty" value="" spellcheck="false">
 	                </div>
 	                <div class="form-group">
-	                  <label>'.$langs->trans("Country").'</label><br>';
+	                  <label>'.$langs->trans("Country").'</label> &nbsp; ';
 					$countryselected = (GETPOSTISSET('country_id') ? GETPOST('country_id', 'aZ09') : $mythirdpartyaccount->country_id);
 					$exclude_country_code = array();
 					if (getDolGlobalString('SELLYOURSAAS_EXCLUDE_COUNTRY_CODES')) {
@@ -111,8 +111,10 @@ print '
 					print $form->select_country($countryselected, 'country_id', '', 0, 'minwidth300', 'code2', 0, 1, 0, $exclude_country_code);
 					print '</div>'."\n";
 
+
 					if (!getDolGlobalInt('SELLYOURSAAS_ONLY_NON_PROFIT_ORGA')) {
-						print '<div class="form-group">
+						// VAT Section
+						print '<div class="form-group"><br>
 	                                  <label>'.$langs->trans("VATIntra").'</label> ';
 						if (! empty($mythirdpartyaccount->tva_assuj) && empty($mythirdpartyaccount->tva_intra)) {
 							print img_warning($langs->trans("Mandatory"), '', 'hideifnonassuj');
@@ -131,11 +133,11 @@ print '
 
 						print '
 							<br>
-		                  <input type="hidden" name="vatassuj_old" value="'.($mythirdpartyaccount->tva_assuj).'">
+		                  <input type="hidden" name="vatassuj_old" value="'.$mythirdpartyaccount->tva_assuj.'">
 		                  <input type="checkbox" style="margin-bottom: 3px;" class="inline-block valignmiddle"'.($mythirdpartyaccount->tva_assuj ? ' checked="checked"' : '').' id="vatassuj" name="vatassuj"> <label for="vatassuj" class="valignmiddle nobold">'.$langs->trans("IHaveAVATID").'</label>
 							<br>
 		                  <input type="hidden" name="vatnumber_old" value="'.$mythirdpartyaccount->tva_intra.'">
-		                  <input type="text" class="input-small quatrevingtpercent hideifnonassuj" value="'.$mythirdpartyaccount->tva_intra.'" name="vatnumber" id="vatnumber" placeholder="'.$placeholderforvat.'">
+		                  <input type="text" class="input-small quatrevingtpercent hideifnonassuj" value="'.$mythirdpartyaccount->tva_intra.'" name="vatnumber" id="vatnumber" spellcheck="false" placeholder="'.$placeholderforvat.'">
 		                    ';
 						print "\n";
 						print '<script>';
@@ -164,6 +166,47 @@ print '
 							}
 							print $s;
 						}
+						print '</div>'."\n";
+
+
+						// ID Prof section
+						$placeholderforprofid = '';
+						$mandatoryprofid = 0;
+						if ($mythirdpartyaccount->country_code == 'FR') {
+							//$placeholderforprofid='Exemple: FR12345678';
+							$mandatoryprofid = 1;
+						} elseif ($mythirdpartyaccount->country_code == 'BE') {
+							//$placeholderforprofid='Exemple: BE12345678';
+						} elseif ($mythirdpartyaccount->country_code == 'ES') {
+							//$placeholderforprofid='Exemple: ES12345678';
+						} else {
+							//$placeholderforprofid=$langs->trans("EnterVATHere");
+						}
+
+						print '<div class="form-group"><br>
+	                                  <label>'.$langs->transcountry("ProfId1Short", $mythirdpartyaccount->country_code).'</label> ';
+						if ($mandatoryprofid && empty($mythirdpartyaccount->idprof1)) {
+							print img_warning($langs->trans("Mandatory"), '', 'hideifnoprof');
+						}
+
+						print '
+							<br>
+		                  <input type="hidden" name="profid_old" value="'.$mythirdpartyaccount->idprof1.'">
+		                  <input type="text" class="input-small quatrevingtpercent" value="'.$mythirdpartyaccount->idprof1.'" name="profid" id="profid" spellcheck="false" placeholder="'.$placeholderforprofid.'">
+		                    ';
+						print "\n";
+						/*
+						print '<script>';
+						print '$( document ).ready(function() {'."\n";
+						print '$("#profid").keyup(function() {'."\n";
+						print "   console.log('We change the profid='+$('#profid').val());\n";
+						print "   if ($('#profid').val() != '')  { $('#profid').prop('checked', true ); }\n";
+						print '});'."\n";
+						print '});'."\n";
+						print '</script>';
+						print "\n";
+						*/
+
 						print '</div>'."\n";
 					}